[CMD] Force recheck on a host also reschedule an host check
When a host is selected and forced to be rechecked, the only command
sent to Nagios was a SCHEDULE_FORCED_HOST_SVC_CHECKS, which rechedule
a check for all services of the host. However the host itself was not
scheduled for recheck.

From now a SCHEDULE_FORCED_HOST_CHECK command is also sent to Nagios
in order to force a recheck of the host.