[QUERY] Fix on scheduled_downtime_depth in service/host status tables
The values that column "scheduled_downtime_depth" can take for an host
or a service is not only either 0 or 1, it is actually 0 or more. More
precisely, ot appears this value corresponds to the number of
scheduled downtime(s) for the current time and there can be more than
one.

For instance, scheduled downtimes on host "srv-www1":
  - today between 2pm and 3pm
  - today between 1pm and 2.30pm

If the current time is between 2pm and 2.30pm, the value of column
"scheduled_downtime_depth" for host "srv-www1" should be 2.

The query used in OpenPOM was based on the fact that value in column
"scheduled_downtime_depth" could be 0 or 1 (and that's all). This
little review handles 0 or more.