[QUERY] Allow to track anything, no matter the status
A comment value begining with an exclamation mark '!' is used to mark
an host or a service as tracked. Tracked entries are always displayed
in OpenPOM, no matter the level selected or state of the host/svc.

To avoid a second select in the query in order to gather comment and
track information about an entry, with use the comment column to
present those two pieces of information.

COMMENT will now take a value between 0 and 3. It must be seen as a
two bits field where:

- bit 0 indicates comment is present (one or more)
- bit 1 indicates that the entry is tracked

This is done that way, using the bit_or MySQL function:

SELECT BIT_OR( IF(substring(comment_data, 1, 1) = '!', 2, 1) )
FROM nagios_commenthistory AS CO
WHERE CO.object_id = S.service_object_id
AND CO.entry_type = 1
AND CO.comment_source = 1
AND CO.deletion_time = '0000-00-00 00:00:00'

Back on PHP in the interface, we can use the fact that any non-zero
value is true in order to test if a bit is set:

- COMMENT & 0x1 will be true if the entry has one or more comments
- COMMENT & 0x2 will be true if the entry has been set to be tracked