[ACTIONS] Reset: send command DEL_ALL_(HOST|SVC)_COMMENTS first
Problem experienced with Nagios 3.2.3 causing Nagios not to be able to
process command DEL_ALL_HOST_COMMENTS correctly. The problem does not
seem to be happening in Nagios 3.3.1. This is how to reproduce it:

( ts=$(date +%s)
  dur=7200
  echo "[$ts] ADD_HOST_COMMENT;blackbird;1;admin;This is a normal comment"
  echo "[$ts] SCHEDULE_HOST_DOWNTIME;blackbird;$ts;$((ts+dur));1;0;$dur;admin;fixed downtime"
) > /var/nagios/rw/command.pipe

Scheduled downtime ID is 47

( ts=$(date +%s)
  echo "[$ts] DEL_HOST_DOWNTIME;47"
  echo "[$ts] DEL_ALL_HOST_COMMENTS;blackbird"
) > /var/nagios/rw/command.pipe

Problem, the normal comment is never deleted. Any subsequent call to
DEL_ALL_HOST_COMMENTS will fail to delete that comment. However
calling DEL_HOST_COMMENT with the comment ID will succeed to delete
that comment.

This only workaround I have been able to find is to first delete all
comments, then cancel the downtime. That way both downtime and normal
comments are deleted and the host downtime is canceled.