[COUNTERS] Fix global counters duplicate rows
Explanation of the problem, using the first subquery to demonstrate:

SELECT
  'current_state_svc'       AS type,
  SS.current_state          AS state,
  COUNT( SS.current_state ) AS total
FROM
       nagios_services AS S
  JOIN nagios_servicestatus AS SS          ON S.service_object_id = SS.service_object_id
  JOIN nagios_service_contactgroups AS SCG ON SCG.service_id = S.service_id
  JOIN nagios_contactgroups AS HCG         ON SCG.contactgroup_object_id = HCG.contactgroup_object_id
  JOIN nagios_contactgroup_members AS CGM  ON CGM.contactgroup_id = HCG.contactgroup_id
  JOIN nagios_contacts AS C                ON C.contact_object_id = CGM.contact_object_id
  JOIN nagios_objects AS O                 ON O.object_id = C.contact_object_id
WHERE
  O.name1 = 'admin'
GROUP BY
  SS.current_state

It's starting first with the services table, then it joins on the
servicestatus table. At this point we still have one row per service.

The problem starts with the join on the service_contactgroups table.
It will cause each row of the service table to be multiplied by the
number of contact groups put on the service and for which the
restricted user belongs to.

It can be reproducted by:
- having more than one contact group on a service ;
- having the restricted user (define_my_user) member of more than one
  of those groups.

For instance say the restricted user is "admin", member of contact
groups "admins" and "test". One of the services has contact_groups set
to both "admins" and "test". The total counters would be doubled.

We fix all sub-queries in the same manner, the exemple above becomes:

  SELECT
    'current_state_svc'       AS type,
    SS.current_state          AS state,
    COUNT( SS.current_state ) AS total
  FROM
         nagios_services AS S
    JOIN nagios_servicestatus AS SS ON S.service_object_id = SS.service_object_id
  WHERE
    'admin' IN (
        SELECT DISTINCT _O.name1
        FROM nagios_services AS _S
        JOIN nagios_service_contactgroups AS _SCG ON _S.service_id = _SCG.service_id
        JOIN nagios_contactgroups AS _CG          ON _SCG.contactgroup_object_id = _CG.contactgroup_object_id
        JOIN nagios_contactgroup_members AS _CGM  ON _CG.contactgroup_id = _CGM.contactgroup_id
        JOIN nagios_contacts AS _C                ON _CGM.contact_object_id = _C.contact_object_id
        JOIN nagios_objects AS _O                 ON _C.contact_object_id = _O.object_id
        WHERE _S.service_id = S.service_id
    )
  GROUP BY
    SS.current_state