Nagios action: Do not use DEL_ALL_HOST_COMMENTS
Nagios command DEL_ALL_HOST_COMMENTS is buggy on Nagios 3.2.3. We
observed that "sometimes" the comments on a host would not be
deleted when sending command DEL_ALL_HOST_COMMENTS to the Nagios
fifo. The problem was resolved in Nagios 3.3.1 but we stick to 3.2.3
for now because of pending bugs.

The workaround is to fetch all individual comments from the NDO
database and send a DEL_HOST_COMMENT command for each of them. This
is also what is done to delete downtimes.