Fix flag indicating disabled checks
This flag was handled in a way that was causing troubles with passive
checks. At first it was added with the assumption that all services
should have either active or passive checks enabled, but not both.
With this change the flag is displayed only if both active and passive
checks are disabled.