Status popin review, ability to reverse $POPIN with CTRL key
Major review and cleanup of the status popin. Added ability to
reverse the "display status popin" setting by holding the CTRL key.
If you felt displaying of the status popin was sometimes boring, that
new feature might worse a try.

If the "display status popin" settings is off, it won't be display
when the mouse is put over a row. However, if you press the CTRL key,
the popin will show.

Similarly, if the "display status popin" settings is turned on, you
can prevent it from being shown when hovering a row by pressing the
CTRL key.