FEATURE/MINOR: query: Add is_ref option for cvar columns
This is useful when the value of the custom variable correspond to
the name of another custom variable. This "referenced" custom variable
must be unique for this feature to work properly.

define host {
    host_name               dummy
    address                 127.0.0.1
    check_command           check-host-nomon
    _SLA_ACME_SEVERITY_1    14400
    _SLA_ACME_SEVERITY_2    43200
    _SLA_ACME_SEVERITY_3    86400
}

define service {
    host_name               mysql1a.acme
    service_description     MYSQL-connect
    ...
    _SLA                    SLA_ACME_SEVERITY_1
}

Now if you want to display the "referenced" SLA value, that is to say,
not SLA_ACME_SEVERITY_1 but 14400 instead, here is how you can define
the column:

$COLUMN_DEFINITION['sla'] = array(
    'cvar'   => 'SLA',
    'is_ref' => true
);

To effectively display the column, it must be added to the
$COLUMN_ENABLED array:

$COLUMN_ENABLED = array(
    [...]
    'sla' => true,
    [...]
);

Conflicts:
	query.php