MINOR: query: Add column option COL_DATA_FIRST
This option is related to the "data" part of a column definition, when
an array of strings is set instead of a single string.

Such "data" properties are automatically created with "cvar" columns:
array("SCVAR_<column>", "HCVAR_<column>")

Meaning... use the value on the service is non-null or else use the
value on the host, even if it's null as well. This behavior now
requires COL_DATA_FIRST to be set in the "opts" property. The default
behavior is now that we concatenate all values, which can be usefull
with COL_MULTI columns. In particular it will allow us to implement
display of hostgroups and servicegroups both in a single column.