MAJOR: query: Optimize query for many custom variable columns
When a custom variable column is used in filtering criteria, use
joins on the host/svc/events union-queries. Otherwise just add
a sub-query in the select wrapper.

There will be another step to optimize the query with a similar
trick but for hgroup/sgroup columns. The joins in union-queries are
useful only if the column are involved in a search.

Conflicts:
	config.php
	index.php
	option.php
	query.php